Breaking out the router tests. Adding tests for dispatcher usage
